Strip metadata added to md files from Notable etc.
Notable creates metadata tags at the top of its files in this format: 

    ---
    tags: [Tags]
    title: Note title
    created: '2019-12-20T00:00:00.000Z'
    modified: '2020-01-20T00:00:00.000Z'
    ---

This patch removes that metadata from the rendered markdown. It should leave pretty much everything else unchanged, as in standard Markdown "---" would be a heading level 2 marker or horizontal rule, and would not appear on the first line.